frontend enhancements
