+++ frontend/index.html
@@ -31,7 +31,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
-      href="https://fonts.googleapis.com/css2?family=Roboto+Flex:wght@300;600&family=IBM+Plex+Mono&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Roboto+Flex:wght@300;600;800&family=IBM+Plex+Mono&display=swap"
       rel="stylesheet"
     />
   </head>
