+++ frontend/src/components/AppLink.vue
@@ -38,7 +38,6 @@ const component = computed(() => (isExternal.value ? "a" : "router-link"));
 .link {
   display: inline-flex;
   align-items: center;
-  justify-content: center;
-  gap: 5px;
+  gap: 0.5em;
 }
 </style>
