+++ frontend/src/styles.css
@@ -1,12 +1,12 @@
 :root {
   /* colors */
-  --theme-light: #f8f3e3;
-  --theme: #cfb87c;
-  --theme-dark: #a87800;
+  --theme-light: hsl(43, 64%, 86%);
+  --theme: hsl(43, 46%, 65%);
+  --theme-dark: hsl(43, 44%, 29%);
   --white: #ffffff;
-  --off-white: #fafafa;
-  --light-gray: #e6e6e6;
-  --gray: #bcbcbc;
+  --off-white: #f8f8f8;
+  --light-gray: #e8e8e8;
+  --gray: #a0a0a0;
   --dark-gray: #434343;
   --off-black: #1a1a1a;
   --black: #000000;
@@ -14,6 +14,7 @@
   /* font weights */
   --regular: 300;
   --bold: 600;
+  --extra-bold: 800;
 
   /* effects */
   --rounded: 3px;
@@ -100,7 +101,8 @@ h4 {
 
 h1 {
   margin: 0;
-  font-size: 1.3rem;
+  font-weight: var(--regular);
+  font-size: 1.2rem;
 }
 
 h2 {
@@ -128,7 +130,7 @@ a {
 }
 
 a:hover {
-  color: var(--off-black);
+  color: currentColor;
 }
 
 b,
@@ -141,7 +143,7 @@ hr {
   flex-shrink: 0;
   width: 100%;
   height: 1px;
-  margin: 20px 0;
+  margin: 5px 0;
   border: none;
   background: var(--gray);
 }
@@ -215,13 +217,6 @@ blockquote {
   align-items: center;
   justify-content: center;
   gap: 10px 20px;
-}
-
-.truncate:not(:focus) {
-  max-width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
 }
 
 .mini-table {
